+++ UI/test/com/kotcrab/vis/ui/test/manual/TestDialogs.java
@@ -22,6 +22,7 @@ import com.kotcrab.vis.ui.util.TableUtils;
 import com.kotcrab.vis.ui.util.Validators;
 import com.kotcrab.vis.ui.util.dialog.ConfirmDialogListener;
 import com.kotcrab.vis.ui.util.dialog.Dialogs;
+import com.kotcrab.vis.ui.util.dialog.Dialogs.DetailsDialog;
 import com.kotcrab.vis.ui.util.dialog.Dialogs.OptionDialogType;
 import com.kotcrab.vis.ui.util.dialog.InputDialogAdapter;
 import com.kotcrab.vis.ui.util.dialog.OptionDialogAdapter;
@@ -44,10 +45,13 @@ public class TestDialogs extends VisWindow {
 		VisTextButton showInputDialogIntOnly = new VisTextButton("show input dialog (int only)");
 		VisTextButton showOptionDialog = new VisTextButton("show option dialog");
 		VisTextButton showConfirmDialog = new VisTextButton("show dialog with custom buttons");
+		VisTextButton showDetailsDialog = new VisTextButton("show details dialog");
+		VisTextButton showDetailsExpandedDialog = new VisTextButton("show expanded details dialog");
 
 		VisTable firstRowTable = new VisTable(true);
 		VisTable secondRowTable = new VisTable(true);
 		VisTable thirdRowTable = new VisTable(true);
+		VisTable fourthRowTable = new VisTable(true);
 
 		firstRowTable.add(showOKMsg);
 		firstRowTable.add(showErrorMsg);
@@ -59,9 +63,13 @@ public class TestDialogs extends VisWindow {
 		thirdRowTable.add(showOptionDialog);
 		thirdRowTable.add(showConfirmDialog);
 
+		fourthRowTable.add(showDetailsDialog);
+		fourthRowTable.add(showDetailsExpandedDialog);
+
 		add(firstRowTable).row();
 		add(secondRowTable).row();
 		add(thirdRowTable).row();
+		add(fourthRowTable).row();
 		padBottom(3);
 
 		showOKMsg.addListener(new ChangeListener() {
@@ -160,8 +168,34 @@ public class TestDialogs extends VisWindow {
 			}
 		});
 
+		final String loremIpsum = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec pretium, justo sed " +
+				"porttitor pretium, justo est porta risus, sed semper elit turpis in leo. Nam aliquam ac augue id congue. " +
+				"Maecenas luctus, tortor et suscipit dictum, augue lectus semper erat, a gravida velit libero vel magna. " +
+				"Nunc vitae fermentum nisl. Cras magna mauris, porttitor dignissim nunc vitae, viverra congue libero. " +
+				"Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos. Duis et " +
+				"suscipit neque, dictum vehicula leo. Sed eu ullamcorper justo, quis sagittis urna. Etiam porttitor ligula " +
+				"a eros rutrum fringilla. Quisque quis ligula at neque aliquet malesuada a ac tellus. ";
+
+		showDetailsDialog.addListener(new ChangeListener() {
+			@Override
+			public void changed (ChangeEvent event, Actor actor) {
+				DetailsDialog dialog = Dialogs.showDetailsDialog(getStage(), "this is a short description text", "details dialog test", loremIpsum);
+				dialog.setCopyDetailsButtonVisible(false);
+				dialog.setWrapDetails(true);
+			}
+		});
+
+		showDetailsExpandedDialog.addListener(new ChangeListener() {
+			@Override
+			public void changed (ChangeEvent event, Actor actor) {
+				DetailsDialog dialog = Dialogs.showDetailsDialog(getStage(), "this is a short description text", "details dialog test", loremIpsum, true);
+				dialog.setCopyDetailsButtonVisible(false);
+				dialog.setWrapDetails(true);
+			}
+		});
+
 		pack();
-		setPosition(25, 146);
+		setPosition(25, 106);
 	}
 
 }